# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sever for Fusion",
        "title": "Fusion API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/follows": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserFollow"
                ],
                "summary": "Create User Followed Streamer",
                "parameters": [
                    {
                        "description": "Follow data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateUserFollowedStreamerRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.UserFollowedStreamerResponse"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/follows/streamers/{streamer_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserFollow"
                ],
                "summary": "List Follows By Streamer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Streamer ID",
                        "name": "streamer_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationResponse-dto_UserFollowedStreamerResponse"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/follows/users/{user_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserFollow"
                ],
                "summary": "List Follows By User",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.PaginationResponse-dto_UserFollowedStreamerResponse"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/follows/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserFollow"
                ],
                "summary": "Get User Followed Streamer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Follow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserFollowedStreamerResponse"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserFollow"
                ],
                "summary": "Update User Followed Streamer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Follow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Follow data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateUserFollowedStreamerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserFollowedStreamerResponse"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserFollow"
                ],
                "summary": "Delete User Followed Streamer",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Follow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateUserFollowedStreamerRequest": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "notification_channel_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "notifications_enabled": {
                    "type": "boolean"
                },
                "streamer_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.PaginationResponse-dto_UserFollowedStreamerResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserFollowedStreamerResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateUserFollowedStreamerRequest": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "notification_channel_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "notifications_enabled": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UserFollowedStreamerResponse": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "notification_channel_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "notifications_enabled": {
                    "type": "boolean"
                },
                "streamer_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}